Added 'job n is already in background' warning to bg.

+++ src/built-ins/cmd_bg.c
@@ -6,7 +6,7 @@
 /*   By: jumanner <jumanner@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 15:36:17 by jumanner          #+#    #+#             */
-/*   Updated: 2023/01/24 14:26:08 by jumanner         ###   ########.fr       */
+/*   Updated: 2023/01/26 10:54:03 by jumanner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,9 @@ int	cmd_bg(char *const *args, t_state *state)
 	job = get_job(args, state);
 	if (!job)
 		return (1);
+	if (job->state == JOB_RUNNING)
+		return (print_error(1,
+				"42sh: bg: job %i already in background\n", job->id));
 	if (!terminal_apply_config(&(state->orig_conf)))
 		return (print_error(1, ERRTEMPLATE_SIMPLE, ERR_TERMIOS_FAIL));
 	if (killpg(job->pids[0], SIGCONT) == -1)
